move class instantiation from dom
